03/03/2025

+++ OurStore/wwwroot/js/ShoppingBag.js
@@ -19,7 +19,7 @@ const viewOneShoppingBagProduct = async (product) => {
     let cloneProduct = template.content.cloneNode(true)
     cloneProduct.querySelector(".image").style.backgroundImage = `url(./Images/${product.image})`
     cloneProduct.querySelector(".itemName").textContent = product.productName
-    cloneProduct.querySelector(".itemNumber").innerText = product.price + ' ₪'
+    cloneProduct.querySelector(".price").innerText = product.price + ' ₪'
     cloneProduct.querySelector(".expandoHeight").addEventListener('click', () => { deleteFromCart(product.productId) })
     document.querySelector("tbody").appendChild(cloneProduct)
 }
@@ -39,7 +39,7 @@ const deleteFromCart = async (productId) => {
 
 
 const getOrderData = () => {
-    const orderDate = "2025-01-12"
+    const orderDate = new Date();
     console.log(orderDate);
     const userId = JSON.parse(sessionStorage.getItem("currentUser"))
     const products = JSON.parse(sessionStorage.getItem("shoppingBag"))
@@ -53,25 +53,32 @@ const getOrderData = () => {
 }
 
 const placeOrder = async () => {
-    const order = getOrderData();
-    try {
-        const responsePost = await fetch('/api/orders', {
-            method: "POST",
-            headers: {
-                "Content-Type": "application/json"
-            },
-            body: JSON.stringify(order) 
-        })
-        const dataPost = await responsePost.json();
-        if (!responsePost.ok)
-            alert("failed to complete the order")
-        else {
-            alert(`order ${dataPost.orderId} completed`)
-            sessionStorage.setItem("shoppingBag", JSON.stringify([]))
-            viewShoppingBagProducts([])
-        }
+    if (sessionStorage.currentUser == null) {
+        const result = confirm("אינך מחובר, להתחברות לחץ אישור")
+        if (result)
+            window.location.href = "home.html"
     }
-    catch (error) {
-        alert(error)
+    else {
+        const order = getOrderData();
+        try {
+            const responsePost = await fetch('/api/orders', {
+                method: "POST",
+                headers: {
+                    "Content-Type": "application/json"
+                },
+                body: JSON.stringify(order)
+            })
+            const dataPost = await responsePost.json();
+            if (!responsePost.ok)
+                alert("failed to complete the order")
+            else {
+                alert(`order ${dataPost.orderId} completed`)
+                sessionStorage.setItem("shoppingBag", JSON.stringify([]))
+                viewShoppingBagProducts([])
+            }
+        }
+        catch (error) {
+            alert(error)
+        }
     }
 }
